Make texblob service method asynchronous
As the method will eventually make a call to a remote server to retrieve
the necessary data (and thus we don't know when it will complete), and
as we cannot block the UI while we wait, we need to coordinate the view
with the response.

This is done using Promises, an asynchronous technique that changes the
signature of the method. Instead of directly returning a texblob object,
a Promise object is returned, which specifies that once ready, and only
then, will a texblob object be returned. Once the texblob is returned,
it is passed into the component's texblob property as before.

In this case, a space is simply left where the data would go whilst
waiting for the data to be retrieved. Once retrieved, the space is
filled.

Here, there is near zero latency as mock data hardcoded into the app is
retrieved. However, this will change once data is retrived from a real
server.