Create text editor component shell
In order to facilitate child routes, give text editor module its own
root component and accompanying router outlet. The root component is
just a shell, and does nothing except provide a router outlet for the
rest of the components in the text editor module. As such, move the
content that was previously in `TextEditorComponent` into the new
component `TextEditorHomeComponent`.