Modify math on confirmation component view
The link http://localhost:4200/shared/1 remains hardcoded. Again, this
eventually should be dynamic, i.e. when a user clicks share, the math on
the page in the text editor is written to the database, and the
resulting id used, e.g. `http://localhost:4200/shared/${newId}`.

However, now the accompanying math displayed on the confirmation page
is hardcoded and will not match the math in the link:

(a) One less call to the API server / simpler code.
(b) The user can click on the link to see the math they just shared.